+++ Tests/BugsnagTests/BugsnagConfigurationTests.m
@@ -585,15 +585,15 @@ NSString * const kBugsnagUserUserId = @"BugsnagUserUserId";
 
 - (void)testMaxBreadcrumb {
     BugsnagConfiguration *config = [[BugsnagConfiguration alloc] initWithApiKey:DUMMY_APIKEY_32CHAR_1];
-    XCTAssertEqual(50, config.maxBreadcrumbs);
+    XCTAssertEqual(100, config.maxBreadcrumbs);
 
     // alter to valid value
     config.maxBreadcrumbs = 10;
     XCTAssertEqual(10, config.maxBreadcrumbs);
 
     // alter to max value
-    config.maxBreadcrumbs = 100;
-    XCTAssertEqual(100, config.maxBreadcrumbs);
+    config.maxBreadcrumbs = 500;
+    XCTAssertEqual(500, config.maxBreadcrumbs);
 
     // alter to min value
     config.maxBreadcrumbs = 0;
@@ -603,8 +603,8 @@ NSString * const kBugsnagUserUserId = @"BugsnagUserUserId";
     config.maxBreadcrumbs = -1;
     XCTAssertEqual(0, config.maxBreadcrumbs);
 
-    // alter to negative value (overflows)
-    config.maxBreadcrumbs = 500;
+    // alter to too large value
+    config.maxBreadcrumbs = 501;
     XCTAssertEqual(0, config.maxBreadcrumbs);
 }
 
@@ -641,7 +641,6 @@ NSString * const kBugsnagUserUserId = @"BugsnagUserUserId";
     XCTAssertNil(config.enabledReleaseStages);
     XCTAssertEqualObjects(@"https://notify.bugsnag.com", config.endpoints.notify);
     XCTAssertEqualObjects(@"https://sessions.bugsnag.com", config.endpoints.sessions);
-    XCTAssertEqual(50, config.maxBreadcrumbs);
     XCTAssertEqual(config.maxStringValueLength, 10000);
     XCTAssertTrue(config.persistUser);
     XCTAssertEqual(1, [config.redactedKeys count]);